# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CountryRequestDTO {

    private Long amountOfCountryToSave;
}